Fixed formatting isue with output. Seperated the solution into two seperated projects, one (class library) for the `Dice` class and the other (Console App) for the user interface.

+++ DiceRoll/Dice.cs
@@ -10,7 +10,7 @@ namespace DiceRoller
 	public class Dice
 	{
 		protected static Random r;
-		public Dice(byte sides)
+		public Dice(int sides)
 		{
 			Sides = sides;
 			byte[] seed = new byte[1];
